+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Each change should be on one line, as GitHub markdown in the release section can
 - `utatime` plugin: fetches lyrics from utatime.com (formerly lyrical-nonsense.com).
 - `utanet` plugin: fetches lyrics from uta-net.com (Uta-Net).
 - `tekstowo` plugin: fetches lyrics from tekstowo.pl.
+- `songtexte` plugin: fetches lyrics, and a German translation if available, from songtexte.com.
 
 ### Changed